requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.9.0
APScheduler==3.10.4
attrs==25.3.0
certifi==2025.4.26
dotenv==0.9.9
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
multidict==6.4.3
propcache==0.3.1
python-dotenv==1.1.0
python-telegram-bot==20.7
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tornado==6.3.3
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
yarl==1.20.0
